refactor(api): v0.20.0 consistency pass on new widgets
The five widgets added in v0.20.0 (`gauge`, `line_gauge`, `breadcrumb`,
`scrollable_with_gutter`, `HighlightRange`) shipped with five different
argument styles. This collapses them onto a single design language so
v0.20 doesn't expose five mismatched mental models for callers to learn:

- `gauge` / `line_gauge` / `breadcrumb` become Egui-style chainable
  builders that auto-render on Drop. `.show()` captures the response.
  Deprecated shims (`gauge_w`, `gauge_colored`, `breadcrumb_sep`,
  `line_gauge_with`) remain for one minor cycle.
- `scrollable_with_gutter` takes `GutterOpts<G>` instead of three
  positional bookkeeping args. `GutterOpts::line_numbers(total, viewport)`
  is the 90% case; users rarely write the closure manually.
- `HighlightRange::single(i)` added as an idiomatic alias for `line(i)`.
- `gauge` / `line_gauge` ratio widened from `f32` to `f64` to match
  `animate_value`, chart APIs, and `progress_bar`. `GaugeResponse.ratio`
  is now `f64`.

Demos:
- `examples/v020_showcase.rs` — single-screen tour of breadcrumb,
  WidthSpec, theme override, animate_bool, on_hover, gauge family,
  named_focus, gutter highlights.
- `examples/v020_regression_panel.rs` — visual regression check
  combining overlay_anchor, modal+tab_trap, chart, sparkline, table,
  scrollable_with_gutter, error_boundary, keymap_help_overlay.

Migration recipes recorded in CHANGELOG `Unreleased > Breaking`. All
existing v0.19 tests + v0.20 tests pass; 11 new tests cover the builder
APIs, deprecation shims, drop-vs-show ergonomics, and f64 precision.

+++ src/widgets/responses.rs
@@ -15,7 +15,7 @@
 ///
 /// ```no_run
 /// # slt::run(|ui: &mut slt::Context| {
-/// let r = ui.breadcrumb(&["Home", "Settings", "Profile"]);
+/// let r = ui.breadcrumb(&["Home", "Settings", "Profile"]).show();
 /// if let Some(i) = r.clicked_segment {
 ///     // navigate to segment `i`
 /// }
@@ -46,13 +46,16 @@ impl std::ops::Deref for BreadcrumbResponse {
 /// Wraps the row-level [`Response`] plus the rendered ratio so callers can
 /// confirm the displayed value (clamped to `0.0..=1.0`). Implements
 /// `Deref<Target = Response>` so `r.hovered` etc. work directly.
+///
+/// Note: `ratio` was widened from `f32` to `f64` in v0.20.0 so the gauge
+/// family aligns with `animate_value`, chart APIs, and `progress_bar`.
 #[derive(Debug, Clone, Default)]
 #[must_use = "GaugeResponse contains interaction state — check .hovered or .ratio"]
 pub struct GaugeResponse {
     /// The row-level interaction response.
     pub response: Response,
     /// The clamped ratio that was rendered (always `0.0..=1.0`).
-    pub ratio: f32,
+    pub ratio: f64,
 }
 
 impl std::ops::Deref for GaugeResponse {
@@ -62,7 +65,13 @@ impl std::ops::Deref for GaugeResponse {
     }
 }
 
-/// Options for [`Context::line_gauge`](crate::Context::line_gauge).
+/// Options struct for the deprecated
+/// [`Context::line_gauge_with`](crate::Context::line_gauge_with) shim.
+///
+/// New code should use the chainable builder
+/// `ui.line_gauge(ratio).label(...).width(...).filled(...).empty(...)`.
+/// This struct stays around for one minor cycle to ease migration of v0.19
+/// call sites that constructed it directly.
 #[derive(Debug, Clone)]
 pub struct LineGaugeOpts {
     /// Fill character. Default: `'━'`.
